+++ TestProjects/GraphicsTestsHDRP/Assets/Tests/CameraOverrideProjectionTest.cs
@@ -33,96 +33,43 @@ public class CameraOverrideProjectionTest : ClusterRendererTestReferenceCamera
     }
 
     [UnityTest]
-    public IEnumerator CompareVanillaAndOverrideProjection([ValueSource("VolumeProfileNames")] string profileName)
+    public IEnumerator CompareReferenceAndOverrideProjection([ValueSource("VolumeProfileNames")] string profileName)
     {
-        InitializeTest();
-        var exceptionHandler = profileName == "FilmGrain" ? () => Debug.LogError("Film grain test requires the LWRP_DEBUG_STATIC_POSTFX scripting symbol to be defined in the Player Settings.") : (Action)null;
-
-        yield return GraphicsTestUtil.PreWarm();
-
-        PostWarmupInit();
-
-        // Set up the projection with the override properties
-        var cameraTransform = m_ReferenceCamera.transform;
-        var projection = m_ClusterRenderer.ProjectionPolicy as CameraOverrideProjection;
-        Assert.That(projection, Is.Not.Null);
-        projection.Overrides = CameraOverrideProjection.OverrideProperty.All;
-        projection.Position = cameraTransform.position;
-        projection.Rotation = cameraTransform.rotation;
-        projection.ProjectionMatrix = Matrix4x4.Perspective(m_ReferenceCamera.fieldOfView, 1, m_ReferenceCamera.nearClipPlane, m_ReferenceCamera.farClipPlane);
-        m_Volume.profile = LoadVolumeProfile(profileName);
-
-        // First we render "vanilla". Use the Reference Camera
-        // to render.
-        m_Camera.gameObject.SetActive(false);
-        m_ClusterRenderer.gameObject.SetActive(false);
-        m_ReferenceCamera.gameObject.SetActive(true);
-
-        yield return GraphicsTestUtil.DoScreenCapture(m_VanillaCapture);
-
-        // Then we activate Cluster Display.
-        m_ClusterRenderer.gameObject.SetActive(true);
-        m_Camera.gameObject.SetActive(true);
-
-        Assert.IsNotNull(m_ClusterRenderer.PresentCamera);
-
-        yield return GraphicsTestUtil.DoScreenCapture(m_ClusterCapture);
-
-        // Even though the cluster camera and the reference camera have
-        // different properties, the OverrideProjection should
-        // make the cluster camera render as if it has the same properties
-        // as the override camera
-        AssertClusterAndVanillaAreSimilar(exceptionHandler);
+        yield return CompareReferenceAndCluster(profileName, () =>
+            {
+                // Set up the projection with the override properties
+                var cameraTransform = m_ReferenceCamera.transform;
+                var projection = m_ClusterRenderer.ProjectionPolicy as CameraOverrideProjection;
+                Assert.That(projection, Is.Not.Null);
+                projection.Overrides = CameraOverrideProjection.OverrideProperty.All;
+                projection.Position = cameraTransform.position;
+                projection.Rotation = cameraTransform.rotation;
+                projection.ProjectionMatrix = Matrix4x4.Perspective(m_ReferenceCamera.fieldOfView, 1, m_ReferenceCamera.nearClipPlane, m_ReferenceCamera.farClipPlane);
+                m_Volume.profile = LoadVolumeProfile(profileName);
+            },
+            profileName == "FilmGrain"
+                ? () => Debug.LogError(
+                    "Film grain test requires the LWRP_DEBUG_STATIC_POSTFX scripting symbol to be defined in the Player Settings.")
+                : null
+        );
     }
 
     [UnityTest]
-    public IEnumerator CompareVanillaAndOverrideProjectionWithOverscan([ValueSource("VolumeProfileOverscanSupportNames")]
-        string profileName)
+    public IEnumerator CompareReferenceAndOverrideProjectionWithOverscan([ValueSource("VolumeProfileOverscanSupportNames")] string profileName)
     {
-        InitializeTest();
-        yield return GraphicsTestUtil.PreWarm();
+        yield return CompareReferenceAndCluster(profileName, () =>
+        {
+            m_ClusterRenderer.Settings.OverScanInPixels = 64;
 
-        PostWarmupInit();
-
-        m_ClusterRenderer.Settings.OverScanInPixels = 64;
-
-        // Set up the projection with the override properties
-        var cameraTransform = m_ReferenceCamera.transform;
-        var projection = m_ClusterRenderer.ProjectionPolicy as CameraOverrideProjection;
-        Assert.That(projection, Is.Not.Null);
-        projection.Overrides = CameraOverrideProjection.OverrideProperty.All;
-        projection.Position = cameraTransform.position;
-        projection.Rotation = cameraTransform.rotation;
-        projection.ProjectionMatrix = Matrix4x4.Perspective(m_ReferenceCamera.fieldOfView, 1, m_ReferenceCamera.nearClipPlane, m_ReferenceCamera.farClipPlane);
-        m_Volume.profile = LoadVolumeProfile(profileName);
-
-
-        // First we render "vanilla". Use the Reference Camera
-        // to render.
-        m_Camera.gameObject.SetActive(false);
-        m_ClusterRenderer.gameObject.SetActive(false);
-        m_ReferenceCamera.gameObject.SetActive(true);
-
-        yield return GraphicsTestUtil.DoScreenCapture(m_VanillaCapture);
-
-        // Then we activate Cluster Display.
-        m_ClusterRenderer.gameObject.SetActive(true);
-        m_Camera.gameObject.SetActive(true);
-
-        Assert.IsNotNull(m_ClusterRenderer.PresentCamera);
-
-        yield return GraphicsTestUtil.DoScreenCapture(m_ClusterCapture);
-
-        // Even though the cluster camera and the reference camera have
-        // different properties, the OverrideProjection should
-        // make the cluster camera render as if it has the same properties
-        // as the override camera
-        AssertClusterAndVanillaAreSimilar(null);
-    }
-
-    [TearDown]
-    public void TearDown()
-    {
-        DisposeTest();
+            // Set up the projection with the override properties
+            var cameraTransform = m_ReferenceCamera.transform;
+            var projection = m_ClusterRenderer.ProjectionPolicy as CameraOverrideProjection;
+            Assert.That(projection, Is.Not.Null);
+            projection.Overrides = CameraOverrideProjection.OverrideProperty.All;
+            projection.Position = cameraTransform.position;
+            projection.Rotation = cameraTransform.rotation;
+            projection.ProjectionMatrix = Matrix4x4.Perspective(m_ReferenceCamera.fieldOfView, 1, m_ReferenceCamera.nearClipPlane, m_ReferenceCamera.farClipPlane);
+            m_Volume.profile = LoadVolumeProfile(profileName);
+        });
     }
 }
